requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyJWT[crypto]==2.9.0
cryptography==43.0.1
requests==2.32.3
gunicorn==22.0.0
